refactor(clinet): change navbar and fab css
