+++ client/src/components/FloatingActionButton/FloatingActionButton.module.css
@@ -10,7 +10,7 @@
   width: 56px;
   height: 56px;
   border-radius: 100%;
-  background: #333;
+  background: #3E50B4;
   border: none;
   outline: none;
   color: #fff;
