+++ client/src/components/Navigation/NavigationItems/NavigationItem/NavigationItem.module.css
@@ -20,7 +20,7 @@
 .NavigationItem a:hover,
 .NavigationItem a:active,
 .NavigationItem a.active {
-  background-color: #8f5c2c;
-  border-bottom: 4px solid #40a4c8;
+  background-color: #3E50B4;
+  border-bottom: 4px solid #fff;
   color: white;
 }
