+++ client/src/components/Navigation/NavigationItems/NavigationItems.module.css
@@ -1,5 +1,5 @@
 .NavigationItems {
-  margin: 0px 10px 0px 0px;
+  margin: 0px;
   padding: 0;
   list-style: none;
   display: flex;
