+++ client/src/components/Navigation/Toolbar/Toolbar.module.css
@@ -4,7 +4,7 @@
   position: fixed;
   top: 0;
   left: 0;
-  background-color: #333;
+  background-color: #121212;
   display: flex;
   justify-content: space-between;
   align-items: center;
